third_party/blink/web_tests/platform/android/external/wpt/pointerevents/compat/pointerevent_touch-action_two-finger_interaction-expected.txt

This is a testharness.js-based test.
[FAIL] touch two-finger pan on 'touch-action: pinch-zoom'
  assert_equals: expected "pointerdown@grey, pointerdown@grey, pointercancel@grey, pointercancel@grey" but got "pointerdown@grey, pointerdown@grey, pointerup@grey, pointerup@grey"
Harness: the test ran to completion.
